Extract all transports to individual modules
