+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "lerna": "2.0.0-beta.36",
+  "lerna": "2.0.0-beta.32",
   "packages": [
     "packages/*"
   ],
