+++ package.json
@@ -10,7 +10,7 @@
     "docsify": "^1.7.4",
     "glob": "^7.0.6",
     "lab": "^10.6.0",
-    "lerna": "2.0.0-beta.36",
+    "lerna": "2.0.0-beta.32",
     "nock": "^9.0.2",
     "nodemon": "^1.9.2",
     "sinon": "^1.17.4",
