+++ packages/datasource-google-sheets/src/plugin.js
@@ -1,60 +1,15 @@
 'use strict'
 
-const Joi = require('joi')
 const GoogleSheetsTransport = require('./google-sheets-transport')
+const configValidator = require('./config-validator')
 
 class RestDatasource {
   register (dashboard, options) {
     dashboard.contributeDatasource('google-sheets', GoogleSheetsTransport)
   }
 
-  static get inlineCredentialsValidation () {
-    return Joi.object().required().keys({
-      type: Joi.string().required().only('service_account').description('Key type'),
-      project_id: Joi.string().required().description('Project name'),
-      private_key_id: Joi.string().required().description('Project name'),
-      private_key: Joi.string().required().description('Project name'),
-      client_email: Joi.string().email().required().description('Project name'),
-      client_id: Joi.string().required().description('Project name'),
-      auth_uri: Joi.string().uri().required().description('Auth Uri'),
-      token_uri: Joi.string().uri().required().description('Token Uri'),
-      auth_provider_x509_cert_url: Joi.string().uri().required().description('Auth provider x509 certificate url'),
-      client_x509_cert_url: Joi.string().uri().required().description('Client x509 certificate url')
-    })
-  }
-
-  static get fileCredentialsValidation () {
-    return Joi.string().required().description('Filesystem path to credentials json, prefixed with "file:"')
-  }
-
-  static get widgetValidation () {
-    return Joi.object().keys({
-      sheet: Joi.string().required().description('Sheet id'),
-      tab: Joi.string().required().description('Tab Name'),
-      columns: this.columnSchema,
-      rows: this.rowSchema,
-      credentials: Joi.alternatives([
-        this.inlineCredentialsValidation,
-        this.fileCredentialsValidation
-      ]).required().description('Service account credentials')
-    }).required().description('Configuration')
-  }
-
-  static get columnSchema () {
-    return Joi.alternatives([
-      Joi.string().required().description('Column heading'),
-      Joi.array().required().description('Array of column headings')
-    ]).required().description('Column or an array of Columns to retrieve')
-  }
-
-  static get rowSchema () {
-    return Joi.alternatives([
-      Joi.number().required().description('Row number'),
-      Joi.object().keys({
-        from: Joi.number().required().description('First row in range'),
-        to: Joi.number().required().description('Last row in range')
-      }).required().description('Range selector')
-    ]).required().description('Row number to retrieve, or object with "from" and "to" row numbers to select a range')
+  get widgetValidation () {
+    return configValidator.widgetValidation
   }
 }
