+++ packages/datasource-google-sheets/src/plugin.spec.js
@@ -4,7 +4,8 @@ const Plugin = require('./plugin')
 
 describe('datasource-google-sheets.plugin', () => {
   it('has register method', (done) => {
-    expect(Plugin).to.include('register')
+    const plugin = new Plugin()
+    expect(plugin.register).to.exist()
     done()
   })
 })
