+++ packages/datasource-random/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@vudash/datasource-random",
+  "version": "1.0.0",
+  "description": "Random Datasource for Vudash",
+  "main": "plugin.js",
+  "scripts": {
+    "test": "../../node_modules/.bin/lab  -I describe,context,it,sinon,fromSrc,fromTest,fromRoot,expect,beforeEach,afterEach,before,after,resource ../../test/unit.lab.js"
+  },
+  "keywords": [
+    "vudash",
+    "transport",
+    "datasource",
+    "chance",
+    "random",
+    "values"
+  ],
+  "author": "Antony Jones",
+  "license": "MIT",
+  "dependencies": {
+    "bluebird": "^3.4.7",
+    "chance": "^1.0.4"
+  }
+}
