+++ packages/datasource-rest/plugin.js
@@ -8,11 +8,7 @@ class RestDatasource {
     dashboard.contributeDatasource('rest', RestTransport)
   }
 
-  static configValidation () {
-
-  }
-
-  widgetValidation (options) {
+  get widgetValidation () {
     return Joi.object({
       url: Joi.string().description('Url to call'),
       method: Joi.string().only('get', 'post', 'put', 'options', 'delete', 'head').description('Http Method'),
