+++ packages/datasource-value/src/value-transport/index.js
@@ -1,0 +1,15 @@
+'use strict'
+
+const Promise = require('bluebird').Promise
+
+class ValueTransport {
+  constructor (options) {
+    this.config = options
+  }
+
+  fetch () {
+    return Promise.resolve(this.config.value)
+  }
+}
+
+module.exports = ValueTransport
