--- packages/transports/README.MD
@@ -1,11 +1,0 @@
-# Vudash Transports
-
-If you aren't a vudash widget developer you probably don't need to worry too much about this module.
-
-The internal module used by many [Vudash](https://npmjs.org/vudash) components to fetch their data.
-
-Vudash Transports knows about a number of data transport mechanisms, such as REST, Static Configuration, Google Sheets API, etc.
-
-What Vudash Transports can do is to make your life as a widget developer much easier, and leave you to work on the appearance and functionality of your widget without worrying about data-sources and configuration validation.
-
-Documentation has moved [to github](https://vudash.com#widget-transports)
