--- packages/transports/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "vudash-transports",
-  "version": "5.2.2",
-  "description": "Various transports for vudash widgets",
-  "main": "index.js",
-  "scripts": {
-    "test": "../../node_modules/.bin/lab -vpDI 'expect,before,beforeEach,after,afterEach,describe,it,fromRoot,fromSrc,resource' ./unit.lab.js"
-  },
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "chance": "^1.0.4",
-    "hoek": "^4.1.0",
-    "joi": "^9.0.4",
-    "request": "^2.75.0",
-    "request-promise": "^4.1.1",
-    "spreadsheet-to-json": "^1.0.5"
-  }
-}
