--- packages/transports/src/transports/value/index.js
@@ -1,22 +1,0 @@
-const Promise = require('bluebird').Promise
-
-const Transport = require('..')
-const Joi = require('joi')
-
-class ValueTransport extends Transport {
-  constructor (descriptor, seed) {
-    super(descriptor)
-  }
-
-  get configValidation () {
-    return Joi.object({
-      value: Joi.any().required().description('Value to return')
-    })
-  }
-
-  fetch () {
-    return Promise.resolve(this.config.value)
-  }
-}
-
-module.exports = ValueTransport
